feat: remove GA
GA may be greate product, but they are too heavy for my site which has
negative impact on performance. I also do not know how to properly use
them to get some better insight into the data. For those reasons I
choose to switch to Plausible analytics, which is payed product but it
is focused on privacy (no data to google), lightweight and provides just
the information I need. Another option was to use Netlify analytics, but
those are more expensive and it seems I have access only to limited
timeperiod dataset.